Full Body BLE Update #2
Making progress in trying the multiple enum sub-states for each button. Haven't completed device # 2 connection in the script yet. Currently sitting at the addition of the Connect2 state, looking to handle the error messages and test whether or not the Button Click Listener activates Scan 2 which detects the device address, and if detected, subsequently triggers the Connect 2 state.

Notes may not make sense, long day today