Public variable OnClick may need to be seperated by breaks to maintain an initial instance for initialization
Combining the button click instance in the original Scan state seems to present errors.

New approach / working theory is that adding a break between Scan and Scan 2, and adding Scan 2 as child object to the OnClick variable will transition into the next scanning state while the subsequent connection states are running for the first device (connectdisconnect in this case).